project structure for multiple optimizers
