+++ src/scalar/mod.rs
@@ -1,0 +1,1 @@
+//! This module contains all minimizers for functions with a single argument.
