+++ src/vector/mod.rs
@@ -1,0 +1,5 @@
+//! This module contains all minimizers for functions with multiple arguments.
+//! 
+mod nelder_mead;
+
+pub use self::nelder_mead::NelderMeadBuilder;
